+++ CHANGE_LOG.md
@@ -25,6 +25,11 @@
 
 ### 版本更新日志
 
+v1.26.0131
+- 增加DisposableManager管理类，方便管理IDisposable对象
+- 增加FadeIn/FadeOut拓展方法，方便UI淡入淡出
+- 优化TimerManager定时器管理类
+  
 v1.25.1108
 - AVProPlayer增加默认透明贴图
 - 修复日志警告不显示bug